+++ Source/Tests/Shell/subtests/b/run.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+set -eu
+
+echo "Test Group B"
+
+cd 01 && ./test.sh && cd ..