+++ Source/Tests/Shell/subtests/run.sh
@@ -3,3 +3,5 @@
 set -eu
 
 cd a && ./run.sh && cd ..
+echo
+cd b && ./run.sh && cd ..