+++ Documentation/Changelog.md
@@ -6,13 +6,15 @@
 
 Release date: not released yet
 
+- GNU toolkit strictness is activated by `GNU_TK_STRICT` environment variable
+
 ### GNU-TK 2025.4
 
 Release date: July 18, 2025
 
 - Use `errexit` and `pipefail` shell options when executing commands
 - Utilize MSYS2 toolkits installed by `msys2/setup-msys2` GitHub action
-- Fixed command arguments handling for Cygwin and MSYS2
+- Fixed handling of command arguments for Cygwin and MSYS2 toolkits
 - Fixed toolkit deduplication
 
 ### GNU-TK 2025.3