+++ Documentation/Changelog.md
@@ -2,6 +2,12 @@
 
 ## 2025
 
+### GNU-TK 2025.6
+
+Release date: not released yet
+
+- Fixed translation of file path arguments for `-f` (`--file`) option
+
 ### GNU-TK 2025.5
 
 Release date: July 28, 2025